code/gpt2-framework/requirements.txt
transformers==4.11.0
torch-1.9.1